rename shader.js
