+++ src/app.js
@@ -52,7 +52,7 @@ require('./3d/immediateMode3D');
 require('./3d/3d_primitives');
 require('./3d/p5.Matrix');
 require('./3d/material');
-require('./3d/shaders');
+require('./3d/shader');
 require('./3d/interaction');
 
 /**
